feat(friend list): merge two files to friend_list one file
